Insertion of Ulisse Wolf mods (#10)
* Insertion of Ulisse Wolf mods

This pull request is to verify the correct modification of the yaml files so that the mods I developed are already compatible with sc4pack

* Fallout Prop Pack Update

Update of yaml files related to Fallout Prop Pack. The task of this update is to verify the correct syntax in writing the yaml files

* Mod maintenance update

This update should contain all the changes necessary for my mods to work properly with SC4Pack. (Note: Freddy Fazbear's Pizza
is still pending update due to the complexity of the mod)

* fix indentation in fallout yaml files

* Freddy Fazbear Pizza Update

Corrected the Freddy Fazbear Pizza file with syntax that should be corected

* Update of problems identified

Carried out correction of problems identified during the control phase

* fix syntax errors

* rename files

* cleanup

---------

Co-authored-by: memo <memo33@users.noreply.github.com>