introduce marker package `config:sc4-edition` to enforce version constraints
See #110.

This allows DLL mods to encode incompatibilities with the macOS version of the game.

For DLLs that require the Windows digital edition, add:
```yaml
dependencies:
- "config:sc4-edition-windows-digital"
```

For DLLs that are compatible with the disc edition as well, add:
```yaml
dependencies:
- "config:sc4-edition"
conflicting:
- "config:sc4-edition-macos"
```

Both cases trigger a dialog to select the version of the game (Windows-digital/Windows-disc/macOS).